Revert "Extract functions related to block configuration Extract functions related to block configuration into a separate module and update imports accordingly."
This reverts commit 7a35e500b338c534f5f35d22a1b06a80b0fbda2c.
