+++ src/lib.rs
@@ -7,13 +7,7 @@ pub mod blocking {
     pub mod ui {
         pub mod spinners;
     }
-    pub mod config {
-        pub mod config;
-    }
     pub mod websites {
-        pub mod hosts_specific_implementation {
-            pub mod hosts;
-        }
         pub mod block_websites;
     }
 }
